// apps/autocad/code-app/src/generated/models/Enmax_autocadreservationsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface Enmax_autocadreservationsBase {
  enmax_autocadreservationid: string;
  enmax_acdnreservationid?: string;
  enmax_acdndrawingcount?: number;
  enmax_acdnsheetsperdrawing?: number;
  enmax_acdnsequencetype?: number;
  enmax_acdnreason?: string;
  enmax_acdnoverride?: boolean;
  enmax_acdnoverridereason?: string;
  enmax_acdnstatus?: number;
  enmax_acdnapprovedon?: string;
  enmax_acdndeclinereason?: string;
  enmax_acdnissuednumbers?: string;
  ownerid: string;
  owneridtype: string;
  statecode: number;
}

export interface Enmax_autocadreservations extends Enmax_autocadreservationsBase {
  createdon?: string;
  modifiedon?: string;
  _enmax_acdnbusiness_value?: string;
  _enmax_acdnasset_value?: string;
  _enmax_acdnunit_value?: string;
  _enmax_acdndomain_value?: string;
  _enmax_acdnsystem_value?: string;
  _enmax_acdnkind_value?: string;
  _enmax_acdnapprover_value?: string;
  _createdby_value?: string;
  _modifiedby_value?: string;
}